Fix build if --with-openssl-legacy-provider set
