+++ ext/openssl/openssl_backend_v1.c
@@ -27,6 +27,22 @@
 #include <openssl/engine.h>
 #endif
 
+void php_openssl_backend_init(void)
+{
+#ifdef LIBRESSL_VERSION_NUMBER
+	OPENSSL_config(NULL);
+	SSL_library_init();
+	OpenSSL_add_all_ciphers();
+	OpenSSL_add_all_digests();
+	OpenSSL_add_all_algorithms();
+	SSL_load_error_strings();
+#else
+	OPENSSL_init_ssl(OPENSSL_INIT_LOAD_CONFIG, NULL);
+#endif
+
+	php_openssl_backend_init_common();
+}
+
 void php_openssl_backend_shutdown(void)
 {
 #ifdef LIBRESSL_VERSION_NUMBER
